Move X icons from entry folder to constants
